Train an Agent to play the SpaceShooter game
In this version of the SpaceShooter game if the yellow spaceship fires and and is inline with the red spaceship the bullet will instantly hit the red spaceship. This change is to make it easier to train the agent since the red spaceship is moving randomly and it can thus not learn to predict where the red spaceship will be.